Add "Door" tag and implement related interactions
The main change in this commit involves the addition of a new "Door" tag into the TagManager in Unity. This is accompanied by the creation of PFInteractionTrigger script metadata file. Elements in existing scenes, such as the WalkDemo scene, have been updated to include these new features and interactions. This update allows those elements to interact with this new "Door" tag, leading to more diverse and interactive elements in the game.